agregado de gradecard

+++ app/src/main/java/com/app/grader/data/database/entitites/GradeEntity.kt
@@ -5,12 +5,12 @@ import androidx.room.Entity
 import androidx.room.PrimaryKey
 
 @Entity(tableName = "grade")
-data class GradeEntity (
+data class GradeEntity(
     @PrimaryKey(autoGenerate = true)
     @ColumnInfo(name = "id") val id: Int = 0,
     @ColumnInfo(name = "course_id") val courseId: Int,
     @ColumnInfo(name = "title") val title: String,
     @ColumnInfo(name = "description") val description: String,
-    @ColumnInfo(name = "grade") val grade: Float,
+    @ColumnInfo(name = "grade") val grade: Double,
     @ColumnInfo(name = "percentage") val percentage: Double,
 )
